Move some of stats to public interface
